Change file structure for lend tab components
