+++ src/components/Charts/Charts.module.less
@@ -24,9 +24,4 @@
 .barChartWrapper {
   width: 144px;
   height: 188px;
-
-  canvas {
-    width: 144px !important;
-    height: 188px !important;
-  }
 }
