+++ src/pages/DashboardPage/components/DashboardLendTab/DashboardLendTab.module.less
@@ -18,3 +18,25 @@
   grid-template-columns: repeat(auto-fill, minmax(162px, 2fr));
   gap: 24px;
 }
+
+.connectedContainer {
+  display: flex;
+  padding: 24px;
+  gap: 24px;
+  width: 100%;
+}
+
+.collectionsSection,
+.additionalContentSection {
+  width: 50%;
+}
+
+.collectionsCardList {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(162px, 2fr));
+  gap: 24px;
+}
+
+.notConnectedContainer {
+  padding: 24px;
+}
