--- src/pages/DashboardPage/components/DashboardLendTab/components/ConnectedContent/ConnectedContent.module.less
@@ -1,17 +1,0 @@
-.container {
-  display: flex;
-  padding: 24px;
-  gap: 24px;
-  width: 100%;
-}
-
-.collectionsSection,
-.additionalContentSection {
-  width: 50%;
-}
-
-.collectionsCardList {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(162px, 2fr));
-  gap: 24px;
-}
