--- src/pages/DashboardPage/components/DashboardLendTab/components/ConnectedContent/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ConnectedContent'
