--- src/pages/DashboardPage/components/DashboardLendTab/components/NotConnectedContent/NotConnectedContent.module.less
@@ -1,9 +1,0 @@
-.container {
-  padding: 24px;
-}
-
-.cardList {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(162px, 2fr));
-  gap: 24px;
-}
