--- src/pages/DashboardPage/components/DashboardLendTab/components/NotConnectedContent/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NotConnectedContent'
